Refactor sync to utilize title count
